+++ central/node/datastore/store/postgres/store.go
@@ -40,6 +40,8 @@ const (
 	// to deal with failures if we just sent it all.  Something to think about as we
 	// proceed and move into more e2e and larger performance testing
 	batchSize = 500
+
+	cursorBatchSize = 50
 )
 
 var (
@@ -68,6 +70,8 @@ type Store interface {
 	GetNodeMetadata(ctx context.Context, id string) (*storage.Node, bool, error)
 	// GetManyNodeMetadata returns nodes without scan/component data.
 	GetManyNodeMetadata(ctx context.Context, ids []string) ([]*storage.Node, []int, error)
+	// WalkByQuery return nodes scoped by the passed query
+	WalkByQuery(ctx context.Context, q *v1.Query, fn func(node *storage.Node) error) error
 }
 
 type storeImpl struct {
@@ -861,6 +865,56 @@ func (s *storeImpl) retryableGetMany(ctx context.Context, ids []string) ([]*stor
 		}
 	}
 	return elems, missingIndices, nil
+}
+
+// WalkByQuery returns the objects specified by the query
+func (s *storeImpl) WalkByQuery(ctx context.Context, q *v1.Query, fn func(node *storage.Node) error) error {
+	defer metrics.SetPostgresOperationDurationTime(time.Now(), ops.WalkByQuery, "Node")
+
+	conn, release, err := s.acquireConn(ctx, ops.WalkByQuery, "Node")
+	if err != nil {
+		return err
+	}
+	defer release()
+
+	tx, err := conn.Begin(ctx)
+	if err != nil {
+		return err
+	}
+	defer func() {
+		if err := tx.Rollback(ctx); err != nil {
+			log.Errorf("error rolling back: %v", err)
+		}
+	}()
+
+	ctxWithTx := postgres.ContextWithTx(ctx, tx)
+	fetcher, closer, err := pgSearch.RunCursorQueryForSchema[storage.Node, *storage.Node](ctxWithTx, pkgSchema.NodesSchema, q, s.db)
+	if err != nil {
+		return err
+	}
+	defer closer()
+	for {
+		rows, err := fetcher(cursorBatchSize)
+		if err != nil {
+			return pgutils.ErrNilIfNoRows(err)
+		}
+		for _, data := range rows {
+			node, exists, err := s.getFullNode(ctx, tx, data.GetId())
+			if err != nil {
+				return err
+			}
+			if !exists {
+				continue
+			}
+			if err := fn(node); err != nil {
+				return err
+			}
+		}
+		if len(rows) != cursorBatchSize {
+			break
+		}
+	}
+	return nil
 }
 
 // GetNodeMetadata gets the node without scan/component data.